bugfix: atomic attribution in cross-validation.
